+++ kernel/tests/test.c
@@ -36,7 +36,7 @@ void force_hardware_irq() {
 
 void force_page_fault() {
 	printf_info("Forcing page fault...");
-	uint32_t* ptr = (uint32_t*)0xA0000000;
+	uintptr_t* ptr = (uintptr_t*)0xA0000000;
 	uint32_t do_fault = *ptr;
 	printf_err("This should never be reached: %d", do_fault);
 }
@@ -52,13 +52,13 @@ void test_interrupts() {
 void test_heap() {
 	printf_info("Testing heap...");
 
-	uint32_t* a = kmalloc(8);
-	uint32_t* b = kmalloc(8);
+	uintptr_t* a = kmalloc(8);
+	uintptr_t* b = kmalloc(8);
 	printf_dbg("a: %x, b: %x", a, b);
 	kfree(a);
 	kfree(b);
 
-	uint32_t* c = kmalloc(12);
+	uintptr_t* c = kmalloc(12);
 	printf_dbg("c: %x", c);
 	kfree(c);
 
@@ -77,7 +77,7 @@ void test_malloc() {
 	uint32_t used = used_mem();
 
 	for (int i = 0; i < 32; i++) {
-		uint32_t* tmp = (uint32_t*)kmalloc(0x1000);
+		uintptr_t* tmp = (uintptr_t*)kmalloc(0x1000);
 		kfree(tmp);
 	}
